Revert "Merge pull request #3 from SwiftPackageIndex/issue-848"
This reverts commit 5ae30ba8d4b159ebe6e8964a50bf1d8936ff28f9, reversing
changes made to 2ecab7e7de26b136e3fc2b2cba2adbc23b6f8674.
